smaller integer field for challenge point

+++ challenges/models.py
@@ -9,7 +9,7 @@ class Challenge(models.Model):
     decription = models.TextField()
     attachment = models.FileField(upload_to="attachments/", blank=True, null=True)
     category = models.ForeignKey('Category')
-    points = models.PositiveIntegerField()
+    points = models.PositiveSmallIntegerField()
 
 class Hint(models.Model):
     text = models.TextField()
